API Reference

SphericalScattering.associatedLegendre

SphericalScattering.derivatieAssociatedLegendre

SphericalScattering.expansion

SphericalScattering.expansion

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.field

SphericalScattering.getFieldType

SphericalScattering.getFieldType

SphericalScattering.getFieldType

SphericalScattering.getFieldType

SphericalScattering.getFieldType

SphericalScattering.getFieldType

SphericalScattering.getFieldType

SphericalScattering.prefac

SphericalScattering.rotate!

SphericalScattering.scatterCoeff

SphericalScattering.scatterCoeff

SphericalScattering.scatterCoeff

SphericalScattering.scatterCoeff

SphericalScattering.scatterCoeff

SphericalScattering.scatterCoeff

SphericalScattering.scatterCoeff

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.scatteredfield

SphericalScattering.translate

SphericalScattering.associatedLegendre — Method

associatedLegendre(n::T, m::T, ϑ::F)

Compute the associated Legendre polynomials according to the definition by Hansen (A1.25) times m divided by sin(ϑ):

m * Pn|m|(cos(ϑ)) / sin(ϑ)

The special values for ϑ ∈ {0, π/2, π} are treated properly.

ϑ ∈ [0, π] assumed

SphericalScattering.derivatieAssociatedLegendre — Method

d Pn|m|(cos(ϑ)) / dϑ

SphericalScattering.expansion — Method

expansion(sphere::PECSphere, excitation::PlaneWave, plm, kr, s, cosϑ, p, n::Int)

Compute functional dependencies of the Mie series for a plane wave travelling in -z direction with polarization in x-direction.

SphericalScattering.expansion — Method

expansion(sphere::PECSphere, excitation::PlaneWave, ka, plm, cosϑ, sinϑ, n::Int)

Compute far-field functional dependencies of the Mie series for a plane wave travelling in -z direction with polarization in x-direction.

SphericalScattering.field — Method

field(excitation::FitzgeraldDipole, point, quantity::FarField; parameter::Parameter=Parameter())

Compute the electric far field radiated by a Fitzgerald dipole at given position and orientation at point.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.field — Method

field(excitation::HertzianDipole, point, quantity::FarField; parameter::Parameter=Parameter())

Compute the electric far field radiated by a Hertzian dipole at given position and orientation at point.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.field — Method

field(excitation::PlaneWave, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field of a plane wave.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.field — Method

field(excitation::PlaneWave, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the magnetic field of a plane wave.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.field — Method

field(excitation::PlaneWave, quantity::Field; parameter::Parameter=Parameter())

Compute the electric field of a plane wave.

SphericalScattering.field — Method

field(excitation::SphericalMode, quantity::Field; parameter::Parameter=Parameter())

Compute the electric field of a spherical mode.

SphericalScattering.field — Method

field(excitation::SphericalModeTE, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field of a TE spherical mode.

SphericalScattering.field — Method

field(excitation::SphericalModeTE, point, quantity::FarField; parameter::Parameter=Parameter())

Compute the electric far-field of a TE spherical mode.

SphericalScattering.field — Method

field(excitation::SphericalModeTE, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field of a TM spherical mode.

SphericalScattering.field — Method

field(excitation::SphericalModeTE, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric far-field of a TM spherical mode.

SphericalScattering.field — Method

field(excitation::Dipole, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field radiated by a Hertzian dipole at given position and orientation at point.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.field — Method

field(excitation::Dipole, point, quantity::MagneticField; parameter::Parameter=Parameter())

Compute the magnetic field radiated by a Hertzian dipole at given position and orientation at point.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.field — Method

field(excitation::Dipole, quantity::Field; parameter::Parameter=Parameter())

Compute the electric field radiated by a magnetic/electric ring current at some position and with some orientation.

SphericalScattering.field — Method

field(excitation::ElectricRingCurrent, r, ϑ, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field radiated by an electric ring current placed in origin at point (r, ϑ)

SphericalScattering.field — Method

field(excitation::ElectricRingCurrent, r, ϑ, quantity::FarField; parameter::Parameter=Parameter())

Compute the (electric) far-field radiated by an electric ring current placed in origin at point (r, ϑ)

SphericalScattering.field — Method

field(excitation::ElectricRingCurrent, r, ϑ, quantity::MagneticField; parameter::Parameter=Parameter())

Compute the magnetic field radiated by an electric ring current placed in origin at point (r, ϑ)

SphericalScattering.field — Method

field(excitation::RingCurrent, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field radiated by a magnetic/electric ring current at some position and with some orientation.

SphericalScattering.getFieldType — Method

getFieldType(excitation::ElectricRingCurrent, quantity::Field)

Do nothing.

SphericalScattering.getFieldType — Method

getFieldType(excitation::FitzgeraldDipole, quantity::Field)

Exchange electric and magnetic field for a magnetic current + apply duality relations.

SphericalScattering.getFieldType — Method

getFieldType(excitation::HertzianDipole, quantity::Field)

Do nothing.

SphericalScattering.getFieldType — Method

getFieldType(excitation::MagneticRingCurrent, quantity::Field)

Exchange electric and magnetic field for a magnetic current.

SphericalScattering.getFieldType — Method

getFieldType(excitation::SphericalMode, quantity::ElectricField)

Do nothing.

SphericalScattering.getFieldType — Method

getFieldType(excitation::SphericalModeTE, quantity::MagneticField)

Exchange TE and TM. Other prefactor: im / ZF.

SphericalScattering.getFieldType — Method

getFieldType(excitation::SphericalModeTM, quantity::MagneticField)

Exchange TE and TM. Other prefactor: im / ZF.

SphericalScattering.prefac — Method

prefac(m::T, n::T)

Prefactor for both types of spherical vector wave functions

SphericalScattering.rotate! — Method

rotate!(points, rotation::SVector{2,T})

Rotate points around x-axis and y-axis away from the z-axis by angles (in radian) specified in rotation[1] and rotation[2].

The points are assumed do be in Cartesian coordinates.

SphericalScattering.scatterCoeff — Method

scatterCoeff(sphere::PECSphere, excitation::ElectricRingCurrent, n::Int, ka)

Compute scattering coefficient for electric ring current.

SphericalScattering.scatterCoeff — Method

scatterCoeff(sphere::PECSphere, excitation::ElectricRingCurrent, n::Int, ka)

Compute scattering coefficient for electric ring current.

SphericalScattering.scatterCoeff — Method

scatterCoeff(sphere::PECSphere, excitation::MagneticRingCurrent, n::Int, ka)

Compute scattering coefficient for magnetic ring current.

SphericalScattering.scatterCoeff — Method

scatterCoeff(sphere::PECSphere, excitation::MagneticRingCurrent, n::Int, ka)

Compute scattering coefficient for magnetic ring current.

SphericalScattering.scatterCoeff — Method

scatterCoeff(sphere::PECSphere, excitation::PlaneWave, n::Int, ka)

Compute scattering coefficients for a plane wave travelling in -z direction with polarization in x-direction.

SphericalScattering.scatterCoeff — Method

scatterCoeff(sphere::PECSphere, excitation::SphericalModeTE, n::Int, ka)

Compute scattering coefficients for a spherical TE mode travelling towards the origin.

SphericalScattering.scatterCoeff — Method

scatterCoeff(sphere::PECSphere, excitation::SphericalModeTM, n::Int, ka)

Compute scattering coefficients for a spherical TM mode travelling towards the origin.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::ElectricRingCurrent, quantity::FarField; parameter::Parameter=Parameter())

Compute the electric far-field scattered by the PEC sphere, where the ring current is placed along the z-axis.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::FitzgeraldDipole, quantity::FarField; parameter::Parameter=Parameter())

Compute the electric far-field scattered by a PEC sphere, where the dipole is placed along the z-axis at z0.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::HertzianDipole, quantity::FarField; parameter::Parameter=Parameter())

Compute the electric far-field scattered by a PEC sphere, where the dipole is placed along the z-axis at z0.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::MagneticRingCurrent, quantity::FarField; parameter::Parameter=Parameter())

Compute the electric far-field scattered by the PEC sphere, where the ring current is placed along the z-axis.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::PlaneWave, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field scattered by a PEC sphere, for an incident plane wave travelling in -z direction with polarization in x-direction.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::PlaneWave, point, quantity::MagneticField; parameter::Parameter=Parameter())

Compute the (electric) far-field scattered by a PEC sphere, for an incident plane wave travelling in -z direction with polarization in x-direction.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::PlaneWave, point, quantity::MagneticField; parameter::Parameter=Parameter())

Compute the magnetic field scattered by a PEC sphere, for an incident plane wave travelling in -z direction with polarization in x-direction.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::PlaneWave, quantity::Field; parameter::Parameter=Parameter())

Compute the electric field scattered by a PEC sphere, for an incident plane wave.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::SphericalModeTE, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field scattered by a PEC sphere, when excited by a spherical mode travelling towards the origin.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::SphericalMode, quantity::Field; parameter::Parameter=Parameter())

Compute the electric field scattered by a dipole at some position and orientation.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::Dipole, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field scattered by a PEC sphere, where the dipole is placed along the z-axis at z0.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::Dipole, point, quantity::MagneticField; parameter::Parameter=Parameter())

Compute the magnetic field scattered by a PEC sphere, where the dipole is placed along the z-axis at z0.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::Dipole, quantity::Field; parameter::Parameter=Parameter())

Compute the electric field scattered by a dipole at some position and orientation.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::RingCurrent, point, quantity::ElectricField; parameter::Parameter=Parameter())

Compute the electric field scattered by the PEC sphere, where the ring current is placed along the z-axis.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

scatteredfield(sphere::PECSphere, excitation::RingCurrent, quantity::MagneticField; parameter::Parameter=Parameter())

Compute the magnetic field scattered by the PEC sphere, where the ring current is placed along the z-axis.

The point and the returned field are in Cartesian coordinates.

SphericalScattering.scatteredfield — Method

field(excitation::ElectricRingCurrent, quantity::Field; parameter::Parameter=Parameter())

Compute the electric field radiated by an electric ring current at some position and orientation

SphericalScattering.translate — Method

translate(points, translation::SVector{3,T})

Translate the points in negative! direction of the translation vector.

All inputs are assumed do be in Cartesian coordinates.